documentation: make MCPlusPlus directory appear in include path
the trick is to first call make install to ensure that header files are
installed and updated, then run doxygen on the installed headers (rather
than on those in the source dir) and on the cpp files in the source dir.
The full include path is then properly stripped by doxygen.